requirements.txt
simplejson==3.16.0
jsonschema==3.2.0
datetime-truncate==1.1.1
parsedatetime==2.4
prometheus-client==0.15.0
python-jenkins==1.7.0
rfc3339==6.2
python-dotenv==0.20.0
influxdb-client==1.32.0